requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
braindecode==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
coverage==7.6.12
cycler==0.12.1
decorator==5.2.1
docker-pycreds==0.4.0
docstring-inheritance==2.2.2
edfio==0.4.5
EDFlib-Python==1.0.8
einops==0.8.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h5py==3.13.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
memory-profiler==0.61.0
mne==1.9.0
mne-bids==0.14
moabb==1.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyriemann==0.6
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyVirtualDisplay==3.0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.12.2
sentry-sdk==2.22.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
skorch==1.1.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.6.0
torchinfo==1.8.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==1.26.20
wandb==0.19.7
wheel==0.45.1
